+++ .devcontainer/node/post-start.sh
@@ -1,12 +1,12 @@
 #!/usr/bin/env bash
 
-# .devcontainer/post-start.sh v.1.2.0
+# .devcontainer/node/post-start.sh v.2.0.0
 
 # This script runs after the Dev Container is started to set up the dev container environment.
 
 set -euo pipefail
 
-echo "Welcome to Matterbridge Plugin Dev Container (post-start.sh)"
+echo "Welcome to Matterbridge Dev Container (post-start.sh)"
 DISTRO=$(awk -F= '/^PRETTY_NAME=/{gsub(/"/, "", $2); print $2}' /etc/os-release)
 CODENAME=$(awk -F= '/^VERSION_CODENAME=/{print $2}' /etc/os-release)
 echo "Distro: $DISTRO ($CODENAME)"
@@ -19,14 +19,15 @@ echo "Date: $(date)"
 echo "Node.js version: $(node -v)"
 echo "Npm version: $(npm -v)"
 echo "Npm cache: $(npm config get cache)"
-echo "Bun version: $(bun -v)"
-echo "Bun global cache: ${HOME}/.bun/install/cache"
 echo ""
 
-echo "1.post-start - Installing the package dependencies..."
+echo "1.post-start - Installing the project dependencies..."
 npm install --no-fund --no-audit
 
-echo "2.post-start - Building the package..."
+echo "2.post-start - Setting node_modules permissions..."
+sudo chown -R node:node ./node_modules
+
+echo "3.post-start - Building the project..."
 npm run build
 
-echo "3.post-start - Post start setup completed!"
+echo "4.post-start - Post start setup completed!"