+++ CHANGELOG.md
@@ -31,6 +31,7 @@ If you like this project and find it useful, please consider giving it a star on
 - [package]: Bump `oxlint` to v.1.78.0.
 - [package]: Bump `oxlint-tsgolint` to v.7.0.2001.
 - [package]: Bump `@types/node` to v.26.2.0.
+- [devcontainer]: Add Dev Container (Bun and Node) v.2.0.0.
 
 <a href="https://www.buymeacoffee.com/luligugithub"><img src="https://matterbridge.io/assets/bmc-button.svg" alt="Buy me a coffee" width="80"></a>